+++ cli/root.go
@@ -8,7 +8,7 @@ import (
 	"github.com/urfave/cli/v3"
 )
 
-var version = "v0.7.1"
+var version = "v0.7.2"
 
 // Execute runs the CLI
 func Execute() {